Breaking out Data page components into separate files
